changed format of control file
control file is now a csv file WITH LINES OF VARIABLE LENGTH. This
means that matlab may have troubles importing it in form of a table as
we did previously. We will modify the functions that make use of the
control file to have a “read line by line” approach instead.