Added try catch for guessing cuts
if at any point of the ellipse fitting or cutting of the perimeter the
routines has an error, the frame will be marked as “error” but, the
analysis will move on to analyze the video.

Error frames will be stored in the control file. In the corrected pupil
perimeter video, they will not be modified, but will be anyways flagged
as “error”.