added frameMask flag to extractPupilPerimeter
this flag will mask in light gray a certain number of pixels at the
border of the videos, effectively creating a frame, whose thickness can
be set by the user.

This is useful for masking out partial black dots in sizeCalibration
videos.

The light gray color was chosen so that the frame mask is “neutral”
both for pupil and glint search.
The flag is set to zero as a default.