glintZone implemented in makeControlFile
This is to allow the routine to discard candidate glints in non
plausible locations. The glint Zone is the circle of allowed locations
for the glint. It is defined by a radius in pixel and a center that can
either be the median glint location in the run or user defined.