Clustering strategy draft
Performance of this approach does not make me happy, mostly because it
is hard to preserve the “couple information” that comes with the fact
that you have 2 glints per frame.

I’ll try a geometrical approach, where I analyze a vector connecting
the 2 glints instead of a single point distribution.

Committing now in case it is a dead end and I need to revert to the
clustering strategy.