Changed blink detection strategy and improved glint patching
After various attempts, it looks like the best performing blink
detection strategy for the 2 glints case is the same as the 1 glint
case, using the mean location obtained from the 2 glints as if they
were a sigle one. This works particularly well in case the 2 glints are
rather close to each other.

The glint patching for the 2 glints case now also adopts the same mean
location strategy and works well in the parallel pool.