potential bug fix
I noticed the behavior that saving videos would sometimes crash, with the error reporting that the size of the frame we're trying to save is different than the size of the rest of the video. This error occured randomly, and would still at least occasionally occur despite the frames all being of the same size. It looks like this may be related to the fact that capture can occur more quickly than figure drawing. Rather than add a pause, I'm trying to force it completely draw with the drawnow command