added new optional parameter, expandPupilRange
This option controls what the routine does when it does not find a circle within the initial range. If true, the routine will expand the pupilRange and look again. If false and no circle is found initially, the routine will move onto the next frame. Consider setting this to false when the routine finds circles that don't correspond to the pupil