implemented new key-value pair 'pickLargestCircle'
This routine uses imfindcircles to identify the likely pupil. However, it can return more than 1 circle that it finds within the pupilMaskFrame. By default it sorts according to most circular, but sometimes the pupil is more likely to be the biggest circle it finds. The default is set to false.