changed behavior for recent frames that lack glint
if a frame within 'recentGlintRange' is a blink (i.e. no glints found), then restore the pupil range to the initial pupil range.

this is to avoid the situation in which the frame just after a blink has partially obscurred pupil -> the routine finds a circle, but it's small -> this sets an artificially small pupil range for the next frames