can now specify 'medianGlintVector' as key-value pair
For the two-glint case, the medianGlintVector is the routine's best guess at the x and y distances between glints. It then uses this guess to figure out which centroids in any given frame are most likely to be the glints.

Sometimes, however, the routine provides a poor guess about the proper x and y distances, and consequently glint tracking fails. This change allows the user to explicitly specifiy what these x and y distances should be.